+++ platformio/platforms/timsp430.py
@@ -23,7 +23,6 @@ class Timsp430Platform(BasePlatform):
         },
 
         "framework-energiamsp430": {
-            "alias": "framework",
             "default": True
         }
     }